move getstatic helpers to a designated folder
