+++ main-site/lib/getStatic/factPage.getStaticProps.ts
@@ -2,9 +2,9 @@ import groq from 'groq';
 import { uniq } from 'lodash-es';
 import { UnwrapPromise } from 'next/dist/lib/coalesced-function';
 import { oneDay } from '~/lib/date';
-import { getFactForumData } from '~/lib/getStatic/getFactForumData';
-import { getRelatedCelebs } from '~/lib/getStatic/getRelatedCelebs';
-import { transformFact } from '~/lib/getStatic/transformFact';
+import { getFactForumData } from '~/lib/getStatic/helpers/getFactForumData';
+import { getRelatedCelebs } from '~/lib/getStatic/helpers/getRelatedCelebs';
+import { transformFact } from '~/lib/getStatic/helpers/transformFact';
 import { Celeb, celebProjection } from '~/lib/groq/celeb.projection';
 import { Fact, factProjection } from '~/lib/groq/fact.projection';
 import { orderOfIssuesGroq } from '~/lib/groq/orderOfIssues.groq';
