+++ main-site/lib/getStatic/helpers/factsDataTransform.ts
