+++ main-site/lib/getStatic/helpers/filterBadWords.ts
