+++ main-site/lib/getStatic/helpers/getFactForumData.ts
@@ -1,5 +1,5 @@
 import { discoursePsqlQuery } from '~/lib/discoursePsqlQuery';
-import { processForumContributorFields } from '~/lib/getStatic/processForumContributorFields';
+import { processForumContributorFields } from '~/lib/getStatic/helpers/processForumContributorFields';
 import { getForumTopicId } from '~/shared/lib/getForumTopicId';
 import { log } from '~/shared/lib/log';
