+++ main-site/lib/getStatic/helpers/groupCelebTags.ts
@@ -1,4 +1,4 @@
-import { sortTags } from '~/lib/getStatic/sortTags';
+import { sortTags } from '~/lib/getStatic/helpers/sortTags';
 import { OrderOfIssues } from '~/lib/groq/orderOfIssues.projection';
 import { Tag } from '~/lib/groq/tag.projection';
 import {
