+++ main-site/lib/getStatic/helpers/processForumContributorFields.ts
