+++ main-site/lib/getStatic/tests/getTagTimeline.test.ts
@@ -1,5 +1,5 @@
-import { rollUp } from '~/lib/getStatic/getTagTimeline';
-import { sortTags } from '~/lib/getStatic/sortTags';
+import { rollUp } from '~/lib/getStatic/helpers/getTagTimeline';
+import { sortTags } from '~/lib/getStatic/helpers/sortTags';
 import { mockOrderOfIssues } from '~/lib/getStatic/tests/mockOrderOfIssues';
 
 test('sortTags to remove dupes', () => {
