+++ main-site/pages/~latest.tsx
@@ -3,7 +3,7 @@ import { CelebImage } from '~/components/CelebImage';
 import { Fact } from '~/components/Fact';
 import { Page } from '~/components/Page';
 import { TitledCard } from '~/components/ui/TitledCard';
-import { transformFact } from '~/lib/getStatic/transformFact';
+import { transformFact } from '~/lib/getStatic/helpers/transformFact';
 import { Fact as TFact, factProjection } from '~/lib/groq/fact.projection';
 import { Picture } from '~/lib/groq/picture.projection';
 import { Link } from '~/lib/Link';
